Move AsmAnalyzer to static context, it no longer uses backend state
741d5892a3 removed the `maxLocalsMaxStackComputed` set of the backend
state. Instead, a flag is set in the MethodNode.

This allows moving `computeMaxLocalsMaxStack` and the `AsmAnalyzer` out
of the backend cake into a static context.
