+++ src/compiler/scala/tools/nsc/backend/jvm/analysis/AliasingAnalyzer.scala
