+++ src/compiler/scala/tools/nsc/backend/jvm/opt/Inliner.scala
@@ -23,7 +23,7 @@ import scala.tools.asm.tree._
 import scala.tools.nsc.backend.jvm.AsmUtils._
 import scala.tools.nsc.backend.jvm.BTypes.InternalName
 import scala.tools.nsc.backend.jvm.BackendReporting._
-import scala.tools.nsc.backend.jvm.analysis.BackendUtils
+import scala.tools.nsc.backend.jvm.analysis.{AsmAnalyzer, BackendUtils}
 import scala.tools.nsc.backend.jvm.opt.BytecodeUtils._
 
 abstract class Inliner {
